Support for Siemens IPC127 MMIO GPIO-driven LEDS
This simply introduces a tiny LED driver based on the MMIO GPIO
registers described in user's manual. It would be nice to find
a generic MMIO GPIO driver since it seems like there's a few that
do this type of thing already (including the one for APU LEDs that
was used as a prototype).

Signed-off-by: Roman Shaposhnik <rvs@apache.org>